added dependency libgsl-dev to package.xml; first implementation of TTC calculation in circle_algorithm.cpp; added helper methods in circle_algorithm

+++ include/ros_collision_detection/ttc_algorithm.h
@@ -25,12 +25,11 @@ typedef struct {
 
 class TTCAlgorithm
 {
-    public:
-        virtual float calculateTTC(
-            const object_motion_t &subject_object_motion,
-            const object_motion_t &perceived_object_motion
-        ) = 0; 
-        
+public:
+    virtual double calculateTTC(
+        const object_motion_t &subject_object_motion,
+        const object_motion_t &perceived_object_motion
+    ) = 0; 
 };
 
 #endif // _TTC_ALGORITM_H_
